# connector/swagger/medco-connector.yml
swagger: "2.0"
info:
  description: "API of the MedCo connector, that orchestrates the query at the MedCo node and provides information about the MedCo network."
  version: "1.0.0"
  title: "MedCo Connector"
  contact:
    email: "medco@epfl.ch"
  license:
    name: "EULA"
    url: "https://raw.githubusercontent.com/ldsec/medco-connector/master/LICENSE"
basePath: "/medco"
...
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
...
security:
  - medco-jwt: []
paths:
...
  /node/analysis/survival/query:
    post :
      tags:
        - "survival-analysis"
      summary: "Send a query to run a survival analysis"
      operationId: "survivalAnalysis"
      parameters:
      - $ref: "#/parameters/survivalAnalysisRequest"

      responses:
        200:
          $ref: "#/responses/survivalAnalysisResponse"
        400:
          $ref : "#/responses/badRequestResponse"
        404:
          $ref : "#/responses/notFoundResponse"
        default:
          $ref: '#/responses/errorResponse'
      security:
        - medco-jwt: ["medco-survival-analysis"]
...
securityDefinitions:
  medco-jwt:
    type: "oauth2"
    flow: "application"
    tokenUrl: "https://medco-demo.epfl.ch/auth"
    description: "MedCo JWT token."
parameters:
...
  survivalAnalysisRequest:
    in: "body"
    description: "User public key, patient list and time codes strings for the survival analysis"
    name: "body"
    required: true
    schema:
      type: "object"
      properties:
        ID:
          type: "string"
          pattern: '^[\w:-]+$'
        userPublicKey:
          type: "string"
          pattern: '^[\w=-]+$'
        cohortName:
          type: "string"
          pattern: '^\w+$'
        subGroupDefinitions:
          type : "array"
          items:
            type: "object"
            properties:
              groupName:
                type: "string"
                pattern: '^\w+$'
              panels:
                type: "array"
                items:
                  $ref: '#/definitions/panel'
          maxItems: 4
        timeLimit:
          type: "integer"
          minimum: 1
        timeGranularity:
          type: "string"
          enum:
            - day
            - week
            - month
            - year
          default: "day"
        startConcept:
          type: "string"
          pattern: '^\/$|^((\/[^\/]+)+\/?)$'
        startModifier:
          type: "object"
          properties:
            modifierKey:
              type: "string"
              pattern: '^((\/[^\/]+)+\/)$'
            appliedPath:
              type: "string"
              pattern: '^((\/[^\/]+)+\/%?)$'
          required:
            - modifierKey
            - appliedPath
        endConcept:
          type: "string"
          pattern: '^\/$|^((\/[^\/]+)+\/?)$'
        endModifier:
          type: "object"
          properties:
            modifierKey:
              type: "string"
              pattern: '^((\/[^\/]+)+\/)$'
            appliedPath:
              type: "string"
              pattern: '^((\/[^\/]+)+\/%?)$'
          required:
            - modifierKey
            - appliedPath
      required:
        - ID
        - userPublicKey
        - cohortName
        - timeLimit
        - timeGranularity
        - startConcept
        - endConcept
responses:
  errorResponse:
    description: "Error response."
    schema:
      type: "object"
      properties:
        message:
          type: "string"
  notFoundResponse:
    description: "Not found."
    schema:
      type: "object"
      properties:
        message:
          type: "string"
...
  badRequestResponse:
    description: "Bad user input in request."
    schema:
      type: "object"
      properties:
        message:
          type: "string"
...
  survivalAnalysisResponse:
    description: "Queried survival analysis"
    schema:
      type: "object"
      properties:
        results:
          type: "array"
          items:
            type: "object"
            properties:
              groupID:
                type: "string"
              initialCount:
                type: "string"
              groupResults:
                type : "array"
                items:
                  type: "object"
                  properties:
                    timepoint:
                      type: "integer"
                    events:
                      type: "object"
                      properties:
                        eventofinterest:
                          type: "string"
                        censoringevent:
                          type: "string"
        timers:
          $ref: "#/definitions/timers"
definitions:
...
  panel:
    type: "object"
    required: [not]
    properties:
      not:
        type: "boolean"
        description: "exclude the i2b2 panel"
      items:
        type: "array"
        description: "i2b2 items (linked by an OR)"
        items:
          type: "object"
          required: [encrypted, queryTerm]
          properties:
            queryTerm:
              type: "string"
              pattern: '^([\w=-]+)$|^((\/[^\/]+)+\/)$'
            operator:
              type: "string"
              enum:
                - EQ
                - NE
                - GT
                - GE
                - LT
                - LE
                - BETWEEN
              description: >
                EQ: equals
                NE: not equals
                GT: greater than
                GE: greater than or equal
                LT: less than
                LE: less than or equal
                BETWEEN: between (value syntax: x and y)
            value:
              type: "string"
              pattern: '^[+-]?([0-9]*[.])?[0-9]+'
            modifier:
              type: "object"
              required: [appliedPath, modifierKey]
              properties:
                appliedPath:
                  type: "string"
                  pattern: '^((\/[^\/]+)+\/%?)$'
                modifierKey:
                  type: "string"
                  pattern: '^((\/[^\/]+)+\/)$'
            encrypted:
              type: "boolean"
      panelTiming:
        $ref: "#/definitions/timing"
...
  timing:
    type: string
    enum:
      - any
      - samevisit
      - sameinstancenum
    default: "any"
...
  timers:
    type: "array"
    items:
      type: "object"
      required: [milliseconds]
      properties:
        name:
          type: "string"
        milliseconds:
          type: "integer"
          format: "int64"